refactor view:  패키지명 변경 ui->view
출입력 받는 패키지명을 ui->view로 변경
